Redesign selection operator
The old code had mostly incorrect behavior that resulted in EANT2 itself
performing poorly on simple problems, such as only having a ~30% success rate
on the XOR problem. The new version is more robust and correct and performs
far better on the XOR problem, with a 100% success rate using on average
1-2 generations.

This is another area where the original paper did not include many details,
but the algorithm implemented here should perform well in general.
Additionally, an option is included to modify the behavior of the selection
process slightly, allowing for future experimentation on different possible
interpretations of the paper.