Change default CMA-ES restart strategy
It seems that two independent CMA-ES restarts are usually used in EANT2, which
is implemented here by the `Local` restart strategy provided by `cmaes`. This
vastly reduces the running time compared to the `BIPOP` strategy. However, one
shortcoming of `Local` is that the population size is fixed to the default,
which is often insufficient for the problem dimensions encountered with larger
networks. This causes a slightly drop in consistency in solving the XOR
problem. In the future, a custom strategy may be implemented to address these
issues, but in the meantime a different strategy may be chosen by the user if
needed.